requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
Django==2.0.13
django-bootstrap4==1.1.1
django-channels==0.7.0
django-filter==2.2.0
django-forms-bootstrap==3.1.0
django-nested-admin==3.3.2
djangorestframework==3.9.4
idna==2.8
Markdown==3.0.1
oauthlib==3.1.0
Pillow==5.4.1
python-monkey-business==1.0.0
pytz==2020.1
requests==2.21.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==1.9.6
stripe==2.49.0
typing==3.7.4.3
urllib3==1.24.3
virtualenv==16.7.9
